Fold the recover helpers into BaseBot's own file
RecoverCallback/RecoverLoop are BaseBot methods guarding BaseBot's own
dispatch and the platform loops built on it - a separate safego.go split
them from the emit/snapshot machinery they belong to. Moved into
core/base.go (tests into core/base_test.go), safego.go deleted; doc and
comment references updated.
