+++ imbot/core/base_test.go
