+++ imbot/core/types.go
@@ -64,7 +64,7 @@ const (
 	// bot's "I crashed" signal to whoever owns its lifecycle: unlike a network
 	// disconnect it must NOT be answered with an in-place reconnect — the
 	// bot's state is suspect and the whole instance should be closed and
-	// rebuilt (see core/safego.go and the host's fatal-panic handling).
+	// rebuilt (see BaseBot.RecoverLoop and the host's fatal-panic handling).
 	ErrPanic ErrorCode = "PANIC"
 )
